Moved useField to hooks folder
