+++ src/hooks/index.ts
@@ -4,3 +4,4 @@ export * from './useFullName';
 export * from './useIsUnmounted';
 export * from './useTimeout';
 export * from './useValidation';
+export * from './useField';
