+++ src/hooks/useField.ts
@@ -1,8 +1,10 @@
-import { useValidation, useFullName, useFormContext, useFieldRegistration, IValidationArgs, IBasicValidationState } from '../../hooks';
-
-import { IFieldComponentFieldProps, IFieldComponentMeta, IBaseFieldProps, TBasicFieldValue, IFieldChangedEvent } from './Field.types';
-import { IValidatedComponentProps } from '../ValidationWrapper';
 import { useMemo, useState, useCallback } from 'react';
+
+import { IBaseFieldProps, IValidatedComponentProps, IFieldComponentFieldProps, IFieldComponentMeta, TBasicFieldValue, IFieldChangedEvent } from '../components';
+import { useFormContext } from './useFormContext';
+import { useFullName } from './useFullName';
+import { useValidation, IValidationArgs, IBasicValidationState } from './useValidation';
+import { useFieldRegistration } from './useFieldRegistration';
 
 export interface IUseFieldProps extends IBaseFieldProps, IValidatedComponentProps { }
